Changed code-block style, added copybutton, increased pagewidth
The pygments style is now manni, however I changed the default colour of imported
packages to a slightly duller blue because the original colour was too jarring.
This change (as well as the change to the page-width), is incorporated in the
newly-linked `_static/custom.css` file, linked to in the conf.py as well. Further
modifications to the custom css can be made by modifying this file. Also, conf.py was
changed to incorporate the new sphinx-copybutton extension (added to requirements.txt).