Added support for LaTeX like dollar-sign math, and extended the epidemiology section
Included basics of modelling epidemics both in the mean-field as well as stochastic regimes, discussed why the residence times in different compartments are exponentially distributed, described the algorithm to add network structure to the simulations, and cleaned up some of the math. The LaTeX-style dollar symbols can now be used, provided you install the sphinx-math-dollar package.